gateway fields sent to separated extension
